+++ src/state/player-location.ts
@@ -31,8 +31,8 @@ prestart(() => {
             }
             packet.playerLocations = memory.diffRecord2Deep(
                 locations,
-                (a, b) => a.pos === b.pos || (a.pos?.z == b.pos?.z && Vec3.equal(a.pos!, b.pos!)),
-                (a: PlayerLocation) => ({ pos: a.pos && Vec3.create(a.pos) })
+                (a, b) => a?.pos?.x == b?.pos?.x && a?.pos?.y == b?.pos?.y,
+                (a: PlayerLocation) => ({ pos: a.pos && Vec2.create(a.pos) })
             )
         },
         set(packet) {